+++ packages/plugin/src/translations/nl/freeform.php
@@ -1174,6 +1174,7 @@ return [
     'Server Checks' => 'Servercontroles',
     'Freeform Stats' => 'Freeform Statistieken',
     'Freeform Checks' => 'Freeform Controles',
+    'Modules' => 'Modules',
 
     // Craft Dashboard Widgets
     'Enter a custom widget title, or leave blank to use default.' => 'Pas de widgettitel aan. Standaard leeg laten.',